+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.8.4] 2021-04-09
+### Added
+* Return if there was an ssl error on each check
+
+### Changed
+* Do not fail on SSL errors but retry on error with ignore ssl errors
+
 ## [0.8.3] 2021-04-05
 ### Changed
 * Always ignore content-type "text/html"